# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amina API
  version: 1.0.0
  description: |
    Image generation and utilities API for the Amina Discord Bot.

    ## Authentication

    All `/v1/*` endpoints require API key authentication via Bearer token.

    ```bash
    curl -H "Authorization: Bearer amina_YOUR_API_KEY" https://api.4mina.app/v1/images/rank-card?...
    ```

    Get your API key at [api.4mina.app/dashboard](https://api.4mina.app/dashboard)

    ## Rate Limiting

    - **60 requests per minute** per API key
    - Rate limit headers included in responses:
      - `X-RateLimit-Limit`: Maximum requests per window
      - `X-RateLimit-Remaining`: Requests remaining
      - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Response Format

    All JSON responses follow this structure:

    ```json
    {
      "success": true,
      "data": { ... },
      "meta": {
        "generatedAt": "2024-01-01T00:00:00.000Z"
      }
    }
    ```

    Image endpoints return SVG directly with `Content-Type: image/svg+xml`.

  contact:
    name: Amina Support
    url: https://discord.gg/4mina

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.4mina.app
    description: Production
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key from dashboard
  parameters:
    imageUrl:
      name: image
      in: query
      required: true
      schema:
        type: string
        format: uri
      description: URL of the image to process
...
  responses:
    SvgImage:
      description: SVG image
      content:
        image/svg+xml:
          schema:
            type: string
...
paths:
...
  /v1/images/generators/changemymind:
    get:
      tags: [Generators]
      summary: '"Change My Mind" meme'
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
          description: Text on the sign
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/generators/drake:
    get:
      tags: [Generators]
      summary: Drake approve/disapprove meme
      parameters:
        - name: top
          in: query
          schema:
            type: string
          description: Text for rejection (top panel)
        - name: bottom
          in: query
          schema:
            type: string
          description: Text for approval (bottom panel)
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/generators/facts:
    get:
      tags: [Generators]
      summary: '"Book of Facts" meme'
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
          description: The "fact" text
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/generators/affect:
    get:
      tags: [Generators]
      summary: '"Does this affect my baby?" meme'
      parameters:
        - $ref: '#/components/parameters/imageUrl'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/generators/beautiful:
    get:
      tags: [Generators]
      summary: '"It''s beautiful" meme'
      parameters:
        - $ref: '#/components/parameters/imageUrl'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'

  # ========== Utilities ==========
...